backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
pydantic-settings>=2.0
websockets>=13.0
httpx>=0.27.0
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20.0
PyJWT>=2.8.0
redis[hiredis]>=5.0.0
openhands-sdk>=0.1.0
openhands-tools>=0.1.0
openhands-workspace>=0.1.0
